seeds/pyproject.toml
[project]
name = "seeds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faker>=33.1.0",
    "pymongo>=4.10.1",
    "python-dotenv>=1.0.1",
]
